# client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/api/sizes": {
      "get": {
        "operationId": "SizesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SizeResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "sizes"
        ]
      },
      "post": {
        "operationId": "SizesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "sizes"
        ]
      }
    },
    "/api/sizes/{id}": {
      "get": {
        "operationId": "SizesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "sizes"
        ]
      },
      "patch": {
        "operationId": "SizesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSizeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "sizes"
        ]
      },
      "delete": {
        "operationId": "SizesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "sizes"
        ]
      }
    },
    "/api/beverages": {
      "get": {
        "operationId": "BeveragesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BeverageResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "beverages"
        ]
      },
      "post": {
        "operationId": "BeveragesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBeverageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeverageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "beverages"
        ]
      }
    },
    "/api/beverages/{id}": {
      "get": {
        "operationId": "BeveragesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeverageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "beverages"
        ]
      },
      "patch": {
        "operationId": "BeveragesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBeverageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeverageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "beverages"
        ]
      },
      "delete": {
        "operationId": "BeveragesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "beverages"
        ]
      }
    },
    "/api/beverages/{id}/sizes": {
      "post": {
        "operationId": "BeveragesController_addSize",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBeverageSizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeverageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "beverages"
        ]
      }
    },
    "/api/beverages/{id}/sizes/{sizeId}": {
      "patch": {
        "operationId": "BeveragesController_updateSizePrice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sizeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBeverageSizePriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeverageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "beverages"
        ]
      },
      "delete": {
        "operationId": "BeveragesController_removeSize",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sizeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "beverages"
        ]
      }
    },
    "/api/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{confirmationNumber}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "confirmationNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ]
      }
    }
  },
  "info": {
    "title": "Lemonade Stand API",
    "description": "Beverage catalog and order processing",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "SizeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Small"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CreateSizeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Small",
            "maxLength": 40
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateSizeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Small",
            "maxLength": 40
          }
        }
      },
      "BeverageSizeOfferingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "size-uuid"
          },
          "name": {
            "type": "string",
            "example": "Small"
          },
          "price": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "id",
          "name",
          "price"
        ]
      },
      "BeverageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Classic Lemonade"
          },
          "sizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BeverageSizeOfferingDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "sizes",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateBeverageDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Classic Lemonade",
            "maxLength": 120
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateBeverageDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Classic Lemonade",
            "maxLength": 120
          }
        }
      },
      "AddBeverageSizeDto": {
        "type": "object",
        "properties": {
          "sizeId": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "example": 2,
            "minimum": 0
          }
        },
        "required": [
          "sizeId",
          "price"
        ]
      },
      "UpdateBeverageSizePriceDto": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "example": 2.5,
            "minimum": 0
          }
        },
        "required": [
          "price"
        ]
      },
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "beverageId": {
            "type": "string",
            "format": "uuid"
          },
          "sizeId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "minimum": 1
          }
        },
        "required": [
          "beverageId",
          "sizeId",
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+1 416 555 0100"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          }
        },
        "required": [
          "customerName",
          "items"
        ]
      },
      "OrderItemResponseDto": {
        "type": "object",
        "properties": {
          "beverageName": {
            "type": "string"
          },
          "sizeName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "type": "number",
            "example": 2
          },
          "lineTotal": {
            "type": "number",
            "example": 4
          }
        },
        "required": [
          "beverageName",
          "sizeName",
          "quantity",
          "unitPrice",
          "lineTotal"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "confirmationNumber": {
            "type": "string",
            "example": "LS-A1B2C3D4"
          },
          "customerName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+1 416 555 0100"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "ada@example.com"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 6
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "confirmationNumber",
          "customerName",
          "phone",
          "email",
          "items",
          "total",
          "createdAt"
        ]
      }
    }
  }
}